Changed the Issues screen to a view issues screen and updated the dashboard to split boats and issues buttons so as to make it easier to understand.
