--- PATPhase2/src/UI/IssuesScreen.java
@@ -1,220 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package UI;
-
-/**
- *
- * @author stuff
- */
-public class IssuesScreen extends javax.swing.JFrame {
-
-    /**
-     * Creates new form IssuesScreen
-     */
-    public IssuesScreen() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        pnlBG = new javax.swing.JPanel();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        listIssues = new javax.swing.JList();
-        pnlFilterOptions = new javax.swing.JPanel();
-        btnFilter = new javax.swing.JButton();
-        chkFixed = new javax.swing.JCheckBox();
-        lblServerity = new javax.swing.JLabel();
-        cboxServerity = new javax.swing.JComboBox();
-        dateFrom = new org.jdatepicker.JDatePicker();
-        lblDateFrom = new javax.swing.JLabel();
-        dateTo = new org.jdatepicker.JDatePicker();
-        lblDateTo = new javax.swing.JLabel();
-        txtBoatName = new javax.swing.JTextField();
-        lblBoatName = new javax.swing.JLabel();
-        pnlAddIssue = new javax.swing.JPanel();
-        pnlFixIssue = new javax.swing.JPanel();
-        jLabel1 = new javax.swing.JLabel();
-
-        setPreferredSize(new java.awt.Dimension(750, 700));
-
-        pnlBG.setBackground(new java.awt.Color(153, 153, 153));
-
-        listIssues.setModel(new javax.swing.AbstractListModel() {
-            String[] strings = { "Item 1", "Item 2", "Item 3", "Item 4", "Item 5" };
-            public int getSize() { return strings.length; }
-            public Object getElementAt(int i) { return strings[i]; }
-        });
-        jScrollPane1.setViewportView(listIssues);
-
-        btnFilter.setText("Filter");
-
-        chkFixed.setText("Fixed");
-
-        lblServerity.setText("Serverity");
-
-        cboxServerity.setModel(new javax.swing.DefaultComboBoxModel(new String[] { "1", "2", "3", " " }));
-
-        lblDateFrom.setText("From");
-
-        lblDateTo.setText("To");
-
-        lblBoatName.setText("Boat Name");
-
-        javax.swing.GroupLayout pnlFilterOptionsLayout = new javax.swing.GroupLayout(pnlFilterOptions);
-        pnlFilterOptions.setLayout(pnlFilterOptionsLayout);
-        pnlFilterOptionsLayout.setHorizontalGroup(
-            pnlFilterOptionsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, pnlFilterOptionsLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(chkFixed)
-                .addGap(18, 18, 18)
-                .addComponent(cboxServerity, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(lblServerity)
-                .addGap(18, 18, 18)
-                .addComponent(dateFrom, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(lblDateFrom)
-                .addGap(18, 18, 18)
-                .addComponent(dateTo, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(lblDateTo)
-                .addGap(18, 18, 18)
-                .addComponent(txtBoatName, javax.swing.GroupLayout.DEFAULT_SIZE, 84, Short.MAX_VALUE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(lblBoatName)
-                .addGap(18, 18, 18)
-                .addComponent(btnFilter)
-                .addContainerGap())
-        );
-        pnlFilterOptionsLayout.setVerticalGroup(
-            pnlFilterOptionsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, pnlFilterOptionsLayout.createSequentialGroup()
-                .addGap(0, 0, 0)
-                .addGroup(pnlFilterOptionsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addGroup(pnlFilterOptionsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(btnFilter, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(chkFixed)
-                        .addComponent(cboxServerity, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(lblServerity)
-                        .addComponent(lblDateFrom)
-                        .addComponent(lblDateTo)
-                        .addComponent(txtBoatName, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(lblBoatName))
-                    .addGroup(pnlFilterOptionsLayout.createSequentialGroup()
-                        .addGap(0, 0, Short.MAX_VALUE)
-                        .addGroup(pnlFilterOptionsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addComponent(dateTo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(dateFrom, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))))
-                .addContainerGap())
-        );
-
-        javax.swing.GroupLayout pnlAddIssueLayout = new javax.swing.GroupLayout(pnlAddIssue);
-        pnlAddIssue.setLayout(pnlAddIssueLayout);
-        pnlAddIssueLayout.setHorizontalGroup(
-            pnlAddIssueLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 0, Short.MAX_VALUE)
-        );
-        pnlAddIssueLayout.setVerticalGroup(
-            pnlAddIssueLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 0, Short.MAX_VALUE)
-        );
-
-        jLabel1.setFont(new java.awt.Font("Tahoma", 0, 18)); // NOI18N
-        jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel1.setText("Fix Issue");
-
-        javax.swing.GroupLayout pnlFixIssueLayout = new javax.swing.GroupLayout(pnlFixIssue);
-        pnlFixIssue.setLayout(pnlFixIssueLayout);
-        pnlFixIssueLayout.setHorizontalGroup(
-            pnlFixIssueLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlFixIssueLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, 216, Short.MAX_VALUE)
-                .addContainerGap())
-        );
-        pnlFixIssueLayout.setVerticalGroup(
-            pnlFixIssueLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlFixIssueLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(368, Short.MAX_VALUE))
-        );
-
-        javax.swing.GroupLayout pnlBGLayout = new javax.swing.GroupLayout(pnlBG);
-        pnlBG.setLayout(pnlBGLayout);
-        pnlBGLayout.setHorizontalGroup(
-            pnlBGLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlBGLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(pnlBGLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(pnlBGLayout.createSequentialGroup()
-                        .addComponent(pnlAddIssue, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addGap(0, 0, 0)
-                        .addComponent(pnlFixIssue, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addComponent(jScrollPane1)
-                    .addComponent(pnlFilterOptions, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        pnlBGLayout.setVerticalGroup(
-            pnlBGLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(pnlBGLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(pnlFilterOptions, javax.swing.GroupLayout.PREFERRED_SIZE, 25, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(pnlBGLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(pnlAddIssue, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(pnlFixIssue, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
-        );
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(pnlBG, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(pnlBG, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * @param args the command line arguments
-     */
-    
-   
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btnFilter;
-    private javax.swing.JComboBox cboxServerity;
-    private javax.swing.JCheckBox chkFixed;
-    private org.jdatepicker.JDatePicker dateFrom;
-    private org.jdatepicker.JDatePicker dateTo;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JLabel lblBoatName;
-    private javax.swing.JLabel lblDateFrom;
-    private javax.swing.JLabel lblDateTo;
-    private javax.swing.JLabel lblServerity;
-    private javax.swing.JList listIssues;
-    private javax.swing.JPanel pnlAddIssue;
-    private javax.swing.JPanel pnlBG;
-    private javax.swing.JPanel pnlFilterOptions;
-    private javax.swing.JPanel pnlFixIssue;
-    private javax.swing.JTextField txtBoatName;
-    // End of variables declaration//GEN-END:variables
-}
